refactor(kube/volumes): check is system
