+++ app/react/kubernetes/volumes/ListView/columns.helper.ts
@@ -1,0 +1,5 @@
+import { createColumnHelper } from '@tanstack/react-table';
+
+import { VolumeViewModel } from './types';
+
+export const helper = createColumnHelper<VolumeViewModel>();
