--- app/react/kubernetes/volumes/ListView/isSystemVolume.ts
@@ -1,9 +1,0 @@
-import KubernetesNamespaceHelper from '@/kubernetes/helpers/namespaceHelper';
-
-import { VolumeViewModel } from './types';
-
-export function isSystemVolume(item: VolumeViewModel) {
-  return KubernetesNamespaceHelper.isSystemNamespace(
-    item.ResourcePool.Namespace.Name
-  );
-}
